Revert "Refactor project structure"
