--- src/pages/Boards/index.jsx
@@ -1,1 +1,0 @@
-// Boards list
