--- src/utils/constants.js
@@ -1,1 +1,0 @@
-export const API_ROOT = 'http://localhost:8017'
