--- src/utils/sorts.js
@@ -1,1 +1,0 @@
-//Sort Columns/Cards
